Jobber apk test

+++ Dockerfile
@@ -1,4 +1,4 @@
-FROM blacklabelops/alpine
+FROM alpine:3.3
 MAINTAINER Steffen Bleul <sbl@blacklabelops.com>
 
 # build parameters
@@ -12,6 +12,7 @@ RUN export JOBBER_HOME=/tmp/jobber && \
     export CONTAINER_USER=jobber_client && \
     export CONTAINER_GROUP=jobber_client && \
     # Install tools
+    apk upgrade --update && \
     apk add --update \
       go \
       git \
@@ -19,22 +20,19 @@ RUN export JOBBER_HOME=/tmp/jobber && \
       wget \
       make && \
     mkdir -p $JOBBER_HOME && \
-    mkdir -p $JOBBER_LIB && \
     # Install Jobber
-    addgroup -g $CONTAINER_GID jobber_client && \
-    adduser -u $CONTAINER_UID -G jobber_client -s /bin/bash -S jobber_client && \
-    cd $JOBBER_LIB && \
-    go get github.com/dshearer/jobber && \
     if  [ "${JOBBER_VERSION}" != "latest" ]; \
-      then \
-        # wget --directory-prefix=/tmp https://github.com/dshearer/jobber/releases/download/v1.1/jobber-${JOBBER_VERSION}-r0.x86_64.apk && \
-        # apk add --allow-untrusted /tmp/jobber-${JOBBER_VERSION}-r0.x86_64.apk ; \
-        cd src/github.com/dshearer/jobber && \
-        git checkout tags/${JOBBER_VERSION} && \
-        cd $JOBBER_LIB ; \
+      then wget --directory-prefix=/tmp https://github.com/dshearer/jobber/releases/download/v1.1/jobber-1.1-r0.x86_64.apk && \
+        apk add --allow-untrusted /tmp/jobber-1.1-r0.x86_64.apk ; \
+      else mkdir -p $JOBBER_LIB && \
+        addgroup -g $CONTAINER_GID jobber_client && \
+        adduser -u $CONTAINER_UID -G jobber_client -s /bin/bash -S jobber_client && \
+        cd $JOBBER_LIB && \
+        go get github.com/dshearer/jobber && \
+        cd $JOBBER_LIB && \
+        make -C src/github.com/dshearer/jobber install DESTDIR=$JOBBER_HOME && \
+        cp $JOBBER_LIB/bin/* /usr/bin ; \
     fi && \
-    make -C src/github.com/dshearer/jobber install DESTDIR=$JOBBER_HOME && \
-    cp $JOBBER_LIB/bin/* /usr/bin && \
     # Install Tini Zombie Reaper And Signal Forwarder
     export TINI_VERSION=0.9.0 && \
     export TINI_SHA=fa23d1e20732501c3bb8eeeca423c89ac80ed452 && \
